refactor: Extract github-related calls
