+++ .gitignore
@@ -99,4 +99,4 @@ __tests__/runner/*
 lib/**/*
 
 lib
-pr-collector/dist
+src/pr-collector/dist
