--- src/pr-collector
@@ -1,1 +1,0 @@
-../pr-collector/src/
