+++ tsconfig.json
@@ -9,5 +9,7 @@
     "esModuleInterop": true,                  /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     "lib": [ "ES2021.String" ]                /* Enable custom `ES2021.String` extension in typescript for `replaceAll` */ 
   },
-  "exclude": ["node_modules", "**/*.test.ts", "pr-collector"]
+  "exclude": ["node_modules", "**/*.test.ts",
+    "src/pr-collector"
+  ]
 }
